requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
yt-dlp==2025.1.26
pydantic==2.8.2
pydantic-settings==2.4.0
google-cloud-storage==2.17.0
python-dotenv==1.0.1
structlog==24.4.0
httpx==0.27.2
